pyproject.toml
[tool.poetry]
name = "berlayar"
version = "0.1.0"
description = ""
authors = ["Enjiao Chen <erniesg@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "1.25.0"
openai = "0.27.8"
astor = "0.8.1"
pytest = "7.1.3"
langchain = "0.0.251"
deeplake = "3.6.15"
python-dotenv = "1.0.0"
sentence-transformers = "2.2.2"
torch = "2.0.1"
pandas = "2.0.3"
tqdm = "4.66.1"
numpy = "1.25.2"
imagebind = "0.0.1.dev0"
protobuf = "3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
